+++ controllers/users.go
@@ -13,4 +13,5 @@ import (
 ******************************************/
 func User(w http.ResponseWriter, r *http.Request) {
 	fmt.Fprintf(w, "Welcome User!")
+
 }
