+++ server/server.go
@@ -13,10 +13,7 @@ import (
 ************************/
 func Start() {
 	router := mux.NewRouter().StrictSlash(true)
-	ConnectMongoDbClient()
 	routes.IntiateRoutes(router)
 	log.Println("Server Started at localhost:8080")
 	http.ListenAndServe(":8080", router)
-
-
 }
